% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooler_to_InteractionSet.R
\name{cooler_to_InteractionSet}
\alias{cooler_to_InteractionSet}
\title{Cooler to InteractionSet}
\usage{
cooler_to_InteractionSet(
  cooler_files,
  gr_window = NULL,
  max_pixels = 1e+07,
  col_data_tibble = NULL,
  col_data_name_field = "sample_name",
  cache_rds = NULL,
  o_write = FALSE,
  o_write_pixels = FALSE
)
}
\arguments{
\item{cooler_files}{Cooler files to be read, one per sample.}

\item{gr_window}{GenomicRanges object with one or more regions to extract. Defaults to NULL, in which case the entire genome will be read.}

\item{max_pixels}{Maximum pixels to extract (fed to read_cooler_hdf5); defaults to 1E7.}

\item{col_data_tibble}{Tibble of column data to merge with the default colData() slot of the IS; defaults to NULL, in which case no data is added.}

\item{col_data_name_field}{Name field of <col_data_tibble> to merge by; should be the names of each cooler file (which will become column names of assay; see also read_cooler_hdf5() behavior).}

\item{cache_rds}{RDS filename under which to cache results (reading full genomes can take a very long time). Defaults to NULL, in which case no cache is produced.}

\item{o_write}{Should existing cache files be over-written? Defaults to FALSE.}

\item{o_write_pixels}{Should an existing pixel cache file (<cache_rds> but with the ".rds" ending replaced with "_pix.tsv") be over-written? Defaults to FALSE, and no effect if a cache file exists and is not being over-written.}
}
\description{
Read and convert a Cooler file (or list of files) into an InteractionSet 
object. Each cooler file will be represented as a single column of each 
assay. 

In addition to the default "counts" assay, this function also includes assays
for the 'count_wgs_nrm' (WGS-normalized counts) column. Additional assays for
filtering (etc.) include:

cnA: Copy number for the first bin (anchor).
cnB: Copy number for the second anchor.
threshold: Expected counts based on same chromosome (distance, logarithmic bins) or different chromosome (median chromosome-chromosome interaction baseline).

All assays represent 2-dimensional data (specific to each pixel and sample).

Default sample-specific data in colData() includes r_name (sample name),
totals (total interactions per sample), and intra_median (intra-chromosome
median per sample).

Default pixel-specific data in rowData() includes bin1_id and bin2_id, as 
well as annotation of intra-chromosomal nature (intra=TRUE), raw distance
between bin midpoints (distance; NA for inter-chromosomal pixels), and 
logarithmic distance bins (dist_bin; NA for inter-chromosomal pixels).
}
