% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_normalize_pixels.R
\name{distance_normalize_pixels}
\alias{distance_normalize_pixels}
\title{Distance-normalize pixels.}
\usage{
distance_normalize_pixels(
  tb_pix_in,
  count_col = "count",
  bins_only = FALSE,
  table_only = FALSE
)
}
\arguments{
\item{tb_pix_in}{Square matrix pixel table from read_cooler_hdf5() (not rotated).}

\item{count_col}{Column of count values to normalize. Defaults to "count".}

\item{bins_only}{Boolean; should only a vector of bin breaks be returned? Defaults to FALSE.}

\item{table_only}{Boolean; should only a table summarizing breaks and baseline values per cell line and chromosome be returned? Defaults to FALSE; superseded by bins_only.}
}
\description{
Given a (non-rotated) pixel table from read_cooler_hdf5(), add columns for
distance-normalized count values (counts to normalize can be provided via
<count_col>, but should be non-logarithmic values, see citation.). Function
will operate in groups based on chromosomes and samples (coolers), and will
separate distances into logarithmic bins (see The Hitchhiker’s guide to Hi-C
analysis: Practical guidelines, Methods, Brian Lajoie, et.al., 2015) which
double in size in series (1MB, 2MB, 4MB, etc.).

Function can return only the break values (bins_only=TRUE), or can return a
tibble with distance-normalized values. Default behavior returns the input
pixel table (tb_pix_in) with new columns "dist" (distance in BP between bins,
NA if bins are on different chromosome), "dist_bin" (the bin in which the
pixel falls), "dist_count" (count column with distance baseline subtracted
and negative values zapped to zero), "dist_base" (baseline of interactions
among pixels in the same distance bin).
}
