% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_matrix.R
\name{downsample_matrix}
\alias{downsample_matrix}
\title{Downsample HiC matrix.}
\usage{
downsample_matrix(
  matrix_in,
  x_bins = 750,
  y_bins = 750,
  return_extra_cols = FALSE
)
}
\description{
Given a HiC matrix (as read by read_matrix_file()), downsample it to have
approximately <bins_X> by <bins_Y> bins. If there fewer than or equal to
that many bins in the original matrix, return the original matrix.
}
