% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axis_labs.R
\name{get_axis_labs}
\alias{get_axis_labs}
\title{Get axis labels}
\usage{
get_axis_labs(pix_in, axis_in = "x", break_num = 5, out_as_vector = TRUE)
}
\arguments{
\item{pix_in}{Matrix tibble in the format returned by read_matrix_file().}

\item{axis_in}{Which axis to return labels for. Defaults to "x", can also be "y". For distance in a rotated matrix, use "y_hic".}

\item{break_num}{Number of breaks to approximate. All chromosome starts must be annotated, however, so if there are more chromosomes than <break_num> it will be effectively ignored.}

\item{out_as_vector}{Should the output table be simplified down to a named vector with the offset start positions as values? Defaults to true.}
}
\description{
Given a tibble returned by read_cooler_hdf5(), determine labels for x/y axes
and return as a table or vector. Specify <break_num> to approximate how many
breaks are required, and all chromosome starts will be included. If more
chromosomes are contained in the data than there are breaks, only chromosome
start points will be annotated, but they all will be regardless of break_num.

Typical format: "Chr1 | 1,234,567 | 106,434,567 | Chr2 | 1,234,567"
}
