% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_figure_dims.R
\name{get_figure_dims}
\alias{get_figure_dims}
\title{Get figuire dimensions}
\usage{
get_figure_dims(
  gr_1 = NULL,
  gr_2 = NULL,
  fig_height = NULL,
  fig_width = NULL,
  max_difference = 20
)
}
\arguments{
\item{gr_1}{GRange region for X-axis. Defaults to entire genome.}

\item{gr_2}{GRange region for Y-axis. Defaults to value of gr_1.}

\item{fig_height}{Desired height of figure (units don't matter).}

\item{fig_width}{Desired width of figure (units don't matter).}

\item{max_difference}{Maximum ratio difference between height and width.}
}
\description{
Given a pair of genomic ranges indicating X and Y axes fore a matrix, return
a vector with appropriate height and width values that reflect the true
dimensions. If one or both GRanges are NULL, the aspect ratio is set to 1.
In the event that one axis is more than <max_difference> times the length of
the other, a warning is displayed and dimensions are returned for the 
<max_difference> ratio. Can be overridden by increasing <max_difference>.
}
