% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionSet_to_pixels.R
\name{InteractionSet_to_pixels}
\alias{InteractionSet_to_pixels}
\title{InteractionSet to pixels}
\usage{
InteractionSet_to_pixels(iset_in)
}
\arguments{
\item{iset_in}{InteractionSet object to convert.}
}
\description{
Given an InteractionSet object, convert it into a pixel table 
compaitible with other hicPac features. All assays will be slotted into one
column of a long-format pixel table, and all assay columns will be assigned 
cooler names in the "cooler" column.
}
