% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyogram_geoms.R
\name{karyogram_geoms}
\alias{karyogram_geoms}
\title{Karyogram geoms}
\usage{
karyogram_geoms(gr_in, axis_in = "x", range_edge)
}
\arguments{
\item{gr_in}{GenomicRanges object representing region to retrive karyotyping data for.}

\item{axis_in}{Axis to configure geoms for, defaults to "x" and can also be "y".}

\item{range_edge}{End of the x or y range as appropriate from which to begin drawing geoms.}
}
\description{
Return geoms for karyotyping strips on either the X or Y axis.
}
