% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_read_assembly_file.R
\name{nlf_read_assembly_file}
\alias{nlf_read_assembly_file}
\title{NeoLoopFinder read assembly file}
\usage{
nlf_read_assembly_file(fl_nm)
}
\arguments{
\item{fl_nm}{Name of file to read in.}
}
\description{
Read in and parse an output assembly file from NeoLoopFinder. Reads file line
by line and applies nfl_parse_assembly() to each line, returning a single
table with assemblies split up with one segment per line.
}
