% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlf_read_loop_file.R
\name{nlf_read_loop_file}
\alias{nlf_read_loop_file}
\title{NeoLoopFinder read loop file}
\usage{
nlf_read_loop_file(file_in = fl_neoloops)
}
\arguments{
\item{file_in}{NeoLoopFinder Neoloop output file}
}
\description{
Given a Neoloop text file output by NeoLoopFinder, parse the file into a
tibble with one row per loop-assembly combination (the same loop can be 
found in multiple lines if it is found in multiple assemblies).
}
