% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_cooler_hdf5.R
\name{pack_cooler_hdf5}
\alias{pack_cooler_hdf5}
\title{Pack Cooler HDF5}
\usage{
pack_cooler_hdf5(
  pixel_tibble,
  cooler_file,
  seq_sizes,
  overwrite_hdf5 = FALSE,
  chunk_size = 1e+07
)
}
\arguments{
\item{pixel_tibble}{Table of data to convert.}

\item{cooler_file}{Cooler filename to create.}

\item{seq_sizes}{Named vector of chromosome sizes to use. If missing, hg38 sizes returned by clugPac::get_seqsizes() will be used.}

\item{overwrite_hdf5}{Should existing cooler files be destroyed? Defaults to FALSE.}

\item{chunk_size}{Size of chunks that pixel values should be split into. Defaults to 1E7.}
}
\description{
Given a table with two-dimensional genome coordinates (specifically columns
for seqnames1, start1, end1, seqnames2, start2, end2), pack this into a
Cooler file. This entails generating bin IDs and an index, but note that
other than chromosome sizes these values are generated based on the table
input: as such it is not guaranteed that bin IDs will match between coolers.

Coordinate columns will be incorporated into bins/indices etc., and all other
columns are treated as columns for the pixel group. If the specified cooler
file already exists that filename is returned with a warning unless
overwrite_hdf5=TRUE, in which case that file will be deleted and a new one
generated in its place.
}
