% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_annotations_2d.R
\name{patchwork_annotations_2d}
\alias{patchwork_annotations_2d}
\title{Patchwork annotations 2D.}
\usage{
patchwork_annotations_2d(
  granges1_in,
  granges2_in,
  regions_tb,
  hic_file,
  gr_segments
)
}
\arguments{
\item{regions_tb}{Table from patchwork_bins(boundaries_only=TRUE) defining the contig in question.}

\item{hic_file}{Cooler file from which to extract bin information if <regions_tb> is NULL; ignored otherwise.}

\item{gr_segments}{GenomicRanges object containing contig segment information to be used if <regions_tb> is NULL; ignored otherwise.}

\item{granges_in}{GenomicRanges object containing annotations to transform.}
}
\description{
Given a GenomicRanges object of genome annotations representing
two coordinates in a patchwork (for example, a translocation), and a region
table from patchwork_bins(boundaries_only=TRUE), subset and clip the contents
of these ranges to identify features which fall within the matrix.


segment's range. Annotates truncated features with "trunc_start" and
"trunc_end" columns, and while start/end coordinates are provided, X and Y
coordinates for plotting are found in "bin_start" and "bin_end" columns,
which have contig-specific values. These values also respect reversal:
features which fall on segments that have been added in the 3'-5' direction
are reversed ("start" and "end" coordinates run backwards), and strand
information is inverted ("+" strand genes are labeld "-"). If <regions_tb>
is not provided, <hic_file> and <gr_segments> can be substituted to execute
patchwork_bins() within the function.
}
