% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_bedgraph.R
\name{patchwork_bedgraph}
\alias{patchwork_bedgraph}
\title{Patchwork BedGraph}
\usage{
patchwork_bedgraph(
  regions_tb,
  bdg_files,
  bdg_names,
  bdg_weights,
  epitope_colors,
  bdg_units = "Pileup",
  raw_alpha = 0.5,
  weighted_alpha = 1,
  num_y_brks = 3
)
}
\arguments{
\item{regions_tb}{Region table, generally the output of patchwork_bins(boundaries_only=TRUE).}

\item{bdg_files}{Filenames of Tabix-indexed BedGraphs. Assumes index file is in place.}

\item{bdg_names}{Names for each BedGraph file, generally their signal contents (i.e. epitope).}

\item{bdg_weights}{Weights to be multiplied against scores in each bedGraph file. Defaults to 1.}

\item{epitope_colors}{Specify colors for each included epitope to be passed to ggplot2's scale_fill_manual(); i.e. (K27Ac="green",K4Me3="blue"). Should be named similar to bdg_files, but doesn't need to be.}

\item{bdg_units}{Y-scale title; defaults to "Pileup."}

\item{raw_alpha}{Alpha (transparency) between 0 and 1 for raw scores; defaults to 0.5.}

\item{weighted_alpha}{Alpha (transparency) between 0 and 1 for weight-multiplied scores; defaults to 1.}

\item{num_y_brks}{Number of Y-axis breaks to be drawn. Defaults to 3.}
}
\description{
Plot signal from BedGraph files within a genome patchwork, respecting
segment orientation. Returns a GGplot object.
}
