% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_bins.R
\name{patchwork_bins}
\alias{patchwork_bins}
\title{Patchwork bins}
\usage{
patchwork_bins(gr_list, hic_file, boundaries_only = TRUE)
}
\arguments{
\item{gr_list}{List of GRanges to combine.}

\item{hic_file}{Cooler filename from which to retrieve bins.}

\item{boundaries_only}{Defaults to TRUE, in which case only bins which bound each segment are retrived. If FALSE, all bins are retrieved.}
}
\description{
Get bins and coordinates for a given contig of DNA starting with a list of
genomic ranges of DNA segments which comprise that contig. Can be named, but
if not each segment will be labeled "reg1", "reg2", etc. Will also append a
"bin_alt" coordinate system which is specific to this contig and is in units
of pixels (the start of the contig is pixel 1, aka bin_alt1_1). Direction of
each segment is indicated by its range's strand ("-" indicates 3' to 5'
direction).
}
