% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_ogm_contigs.R
\name{patchwork_ogm_contigs}
\alias{patchwork_ogm_contigs}
\title{Patchwork OGM contigs}
\usage{
patchwork_ogm_contigs(
  tb_ogm_in,
  regions_tb = NULL,
  hic_file = NULL,
  gr_segments = NULL,
  label_contigs = FALSE,
  label_xmaps = FALSE,
  contig_size = 3,
  xmap_size = 2,
  junctions_only = TRUE,
  return_tables_only = FALSE
)
}
\arguments{
\item{tb_ogm_in}{Tibble of OGM contig/XMAP/CMAP data, typically produced by clugPac::parse_ogm_reads().}

\item{regions_tb}{Tibble of genomic regions ordered from 5' to 3', including strand information to denote orientation. Typically generated by hicPac::patchwork_regions().}

\item{hic_file}{Cooler file which can be used to extract HiC bin locations if regions_tb is not provided.}

\item{gr_segments}{GRanges object of genomic regions stitched, listed in order of 5' to 3' with strand indicating orientation. Only used if regions_tb is not provided.}

\item{label_contigs}{Boolean; should contig IDs be included? Defaults to FALSE.}

\item{label_xmaps}{Boolean; should XMAP IDs be included? Defaults to FALSE.}

\item{contig_size}{Size of contig labels. Defaults to 4.}

\item{xmap_size}{Size of XMAP labels. Defaults to 3.}

\item{junctions_only}{Boolean; should contigs with no XMAPs between stitched regions be omitted? Defaults to TRUE.}

\item{return_tables_only}{Boolean; should plotting be skipped and tables provided instead? Defaults to FALSE, otherwise three tables are returned with contig, xmap, and contig data.}
}
\description{
Given a table of OGM contigs (typically from clugPac::parse_ogm_reads()) and
a table of HiC regions compatible with other patchwork_* functions (produced
by hicPac::patchwork_bins()), split OGM XMAP/CMAP/contig annotations into
contig panels compatible with other hicPac::patchwork_* functions and return
ggplot object illustrating them. If needed,a Cooler file and list of genomic
ranges can be provided in lieu of regions_tb, but these will just be used to
generate regions_tb anyway.

Function can also return a formatted table with region- and strand-oriented
bin ID coordinates (bin_start/end) and y-values based on contig IDs. Intended
for use with HiC patchwork_* functions, so it assumes that the regions_tb
file is formatted as such, but in principle this table can be generated
independently for use elsewhere.
}
