% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_rot45.R
\name{patchwork_rot45}
\alias{patchwork_rot45}
\title{Patchwork rotate 45 degrees}
\usage{
patchwork_rot45(mtx_in)
}
\arguments{
\item{mtx_in}{Patchwork matrix produced by patchwork_matrix() to be rotated.}
}
\description{
Given a matrix of patchwork pixels (typically the output of 
patchwork_matrix()), convert all bins to pixels and rotate their coordinates
by 45 degrees about the left-most pixels in the contig in question. Produces
a table 4x longer than its input (one line per corner of each pixel), and 
produces <x_coords>, <y_coords>, and <pix_id> columns that can be used to 
plot in GGPlot2, i.e. ggplot(mtx,aes(x=x_coords,y=y_coords,group=pix_id)).
}
