% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_rotate_annotations.R
\name{patchwork_rotate_annotations}
\alias{patchwork_rotate_annotations}
\title{Rotate patchwork annotations 45 degrees.}
\usage{
patchwork_rotate_annotations(
  features_tb,
  regions_tb = NULL,
  gr_list = NULL,
  hic_file = NULL,
  seqnames1_field = "seqnames1",
  start1_field = "start1",
  seqnames2_field = "seqnames2",
  start2_field = "start2",
  name_field,
  id_fields
)
}
\arguments{
\item{features_tb}{Table of genomic features which has at minimum two seqnames and two start fields.}

\item{regions_tb}{Regions table (output of hicPac::patchwork_bins(boundaries_only=TRUE)).}

\item{gr_list}{Contig segment list from which to retrieve segment coordinates if <regions_tb> is not provided.}

\item{hic_file}{Cooler file from which to retrieve bin data if <regions_tb> is not provided.}

\item{seqnames1_field}{Name of column for chromosome in first annotation, defaults to "seqnames1".}

\item{start1_field}{Name of column for BP coordinate in first annotation, defaults to "start1".}

\item{seqnames2_field}{Name of column for chromosome in second annotation, defaults to "seqnames2".}

\item{start2_field}{Name of column for BP coordinate in second annotation, defaults to "start2".}

\item{name_field}{Optional; name of column containing name/id data that will be carried over into the output. If not provided, annotations will be labeled "annot1" etc.}

\item{id_fields}{Optional; names of columns containing ID data that is identical for both variants and should be included in the output table (cell line, SV type, etc.).}
}
\description{
Given a table of annotations specific to a patchwork (generally
the output of patchwork_annotations()), calculate "rotated" version which
include X1 and X2 coordinates and a Y coordinate denoting their intersection
in a rotated HiC matrix. Appends columns x1 (bin of first annotation), x2
(mid point between first and second annotation), x3 (bin of second
annotation, and y2 (point on Y-axis where junction should reside; y1 and y3
in this context would both be zero). Annotations can have a name column
(specified via <name_field>), but otherwise they will be assigned names in
the format of "annot_1", "annot_2", etc.

NOTE: IDs may not be unique in cases where the same genome segment is
included multiple times; these are allowed with this function, but output
tables may be confusing when designing plots if you forget this detail.
}
