% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_yaxis.R
\name{patchwork_yaxis}
\alias{patchwork_yaxis}
\title{Patchwork Y-axis}
\usage{
patchwork_yaxis(tb_region, break_num = 4)
}
\arguments{
\item{tb_region}{Table defining the regions of the contig, generally the output of patchwork_bins(boundaries_only=TRUE).}

\item{break_num}{Number of breaks to include on the axis; defaults to 4.}
}
\description{
Given a table defining a given patchwork contig (generally the 
output of patchwork_bins(boundaries_only=TRUE)), define a Y-axis and breaks
which are more descriptive of distance between points on the X-axis.
}
