% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hic.R
\name{plot_hic}
\alias{plot_hic}
\title{Plot HiC}
\usage{
plot_hic(matrix_tbl, reflect_xy = FALSE)
}
\arguments{
\item{matrix_tbl}{Tibble returned by read_hic_matrix().}

\item{reflect_xy}{Draw both (i,j) and (j,i) tiles (reflect matrix across diagonal i=j). Defaults to FALSE.}
}
\description{
Plot a standard HiC interaction matrix (input format should be that returned
by read_hic_matrix()).
}
