% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hic_histogram.R
\name{plot_hic_histogram}
\alias{plot_hic_histogram}
\title{Plot HiC histogram}
\usage{
plot_hic_histogram(
  gr_trap,
  gr_window = "chrom",
  cooler_fls,
  o_write = FALSE,
  return_plot = TRUE,
  wgs_nrm = TRUE,
  log_scale = FALSE,
  max_pix = 5e+06
)
}
\arguments{
\item{gr_trap}{Genomic region to "trap," i.e. what region to detect interactions with.}

\item{gr_window}{Genomic region within which to search for interactions. Can be a specific GRange, "chrom" (search the trap region's chromosome), or "genome" (search the entire genome).}

\item{cooler_fls}{Which cooler files to use when determining bins and searching for interactions.}

\item{o_write}{Should existing <cache_tsv> files be regenerated? Defaults to FALSE.}

\item{wgs_nrm}{Should WGS-normalized counts be used? Defaults to TRUE.}

\item{retur_plot}{Should a plot be returned? Defaults to TRUE, and if FALSE a list of tables will be returned for use elsewhere.}

\item{log_scales}{Should a log10 scale be sued? Defaults to FALSE.}
}
\description{
Detect interactions in the genome with a specified ("trap") region and
generate a histogram for each bin. Defaults to searching for interactions
between the trap region and the rest of its chromosome (<gr_window>="chrom"),
but can also search the entire genome (<gr_window>="genome") or a spcific
contiguous genomic region (<gr_window>=GRange(...)). Bins are cached to a TSV
(<cache_tsv>) rather than the standard pixel file for read_hdf5_*() functions
and can be overwritten with <o_write>=TRUE. Function can also bypass the plot
output and return a list of tables for generating a similar plot: hist_values
contains the histogram, trap describes the trap region, and x_labs contains
X-axis labels.
}
