% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hic_loops.R
\name{plot_hic_loops}
\alias{plot_hic_loops}
\title{Plot HiC loops}
\usage{
plot_hic_loops(
  pixel_tibble,
  gr_trap = NULL,
  gr_window = NULL,
  invert_y = TRUE,
  color_col = NULL,
  color_col_name = NULL,
  alpha_col = NULL,
  alpha_col_name = NULL,
  default_color = "black",
  default_alpha = 1
)
}
\arguments{
\item{pixel_tibble}{Tibble containing pixel data (not rotated) with the score values of interest.}

\item{gr_trap}{GRanges object of trap regions to filter by. Defaults to NULL, in which case all loops are shown.}

\item{gr_window}{GRanges object denoting the x-range window to be drawn. Defaults to NULL, in which case the window will be bounded by the range of X-values in the data.}

\item{invert_y}{Should the Y-axis be inverted (loops open upward)? Defaults to TRUE.}

\item{color_col}{Column name to which color values should be tied. Defaults to "log10_counts".}

\item{color_col_name}{Name of color column to be used on the legend. Defaults to "-log10(Count)".}

\item{alpha_col}{Column name to which alpha values should be tied. Defaults to NULL.}

\item{alpha_col_name}{Name of alpha column to be used on the legend. Defaults to NULL.}

\item{default_color}{Default color in the event no color scale is used. Defaults to "black".}

\item{default_alpha}{Default alpha values in the event no alpha scale is used. Defaults to 1.}
}
\description{
Given a pixel tibble (non-rotated) with at least one column with
score values, draw geom_curve()s between pixel bins with color and/or alpha
values tied to that score column. If 'trap' regions are provided via gr_trap,
illustrated loops will be restricted to those which start/end in one of those
trap windows. For instance, a 10kb trap centered on the promoter region for a
gene will restrict loops to those which start/end on the gene promoter region
and thus may indicate promoter/enhancer loops.

Note that loops are only drawn if both ends are within the x-axis range; no
loops to outside regions are shown. If x_rng is specified, loops to bins off
of the x-range will be dropped. Loops cannot be drawn between the same bin,
so pixels on the diagonal will not be drawn (distance must be greater than
zero).

Also note that color_col_name and alpha_col_name default to NULL, in which
case color and alpha scales (respectively) will be drawn without a legend,
but the scales will be unaffected.
}
