% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hic_matrix.R
\name{plot_hic_matrix}
\alias{plot_hic_matrix}
\title{Plot HiC matrix}
\usage{
plot_hic_matrix(
  gr_1 = NULL,
  gr_2 = NULL,
  silent = FALSE,
  file_cool,
  tb_pixel = NULL,
  plot_bottom_triangle = TRUE,
  max_pixel_count = 6250000,
  score_col = "log10_count",
  fill_low = "#132B43",
  fill_high = "#56B1F7"
)
}
\arguments{
\item{gr_1}{GRange region for X-axis. Defaults to entire genome.}

\item{gr_2}{GRange region for Y-axis. Defaults to value of gr_1.}

\item{silent}{Boolean, defaults to FALSE. Should the function suppress messages about fetching / processing time for read_cooler_hdf5()?}

\item{file_cool}{Cool file to plot.}

\item{tb_pixel}{Tibble formatted from call to read_cooler_hdf5(), can include additional columns but must have basics.}

\item{plot_bottom_triangle}{Should the bottom reflection of the top triangle be included? Defaults to TRUE.}

\item{max_pixel_count}{Maximum number of pixels to plot, i.e. max_pixels argument for read_cooler_hdf5.}

\item{score_col}{Column in pixels tibble to plot within the heatmap, defaults to "log10_count".}

\item{fill_low}{Color of low values (defaults to black).}

\item{fill_high}{Color of high values (defaults to cyan-ish).}
}
\description{
Basic plot of HiC matrix in a region. Accepts a cooler file name and GRanges
for X and Y axes, but can also accept a tibble (<tb_pixel>) that has been
pre-formatted (generally by a call to read_cooler_hdf5()). In such a case,
GRange and <file_cool> arguments are ignored and their values are inferred
from the input tibble.
}
