% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hic_matrix_to_file.R
\name{plot_hic_matrix_to_file}
\alias{plot_hic_matrix_to_file}
\title{Plot HiC matrix to file.}
\usage{
plot_hic_matrix_to_file(
  gr_1 = NULL,
  gr_2 = NULL,
  silent = FALSE,
  file_cool = NULL,
  tb_pixel = NULL,
  kary_width = 0.025,
  max_pixel_count = 6250000,
  score_col = "log10_count",
  file_output = NULL,
  fig_width = NULL,
  fig_height = NULL,
  fig_units = "in",
  fig_device = "png",
  fig_dpi = 300,
  overwrite_cache = FALSE
)
}
\arguments{
\item{gr_1}{GRange region for X-axis. Defaults to entire genome.}

\item{gr_2}{GRange region for Y-axis. Defaults to value of gr_1.}

\item{silent}{Boolean, defaults to FALSE. Should the function suppress messages about fetching / processing time for read_cooler_hdf5()?}

\item{file_cool}{Cool file to plot.}

\item{tb_pixel}{Tibble formatted from call to read_cooler_hdf5(), can include additional columns but must have basics.}

\item{kary_width}{Width of karyogram in fraction of plot height/width; defaults to 0.25 (2.5% of the total plot height and width are added as a karyogram).}

\item{max_pixel_count}{Maximum number of pixels to plot, i.e. max_pixels argument for read_cooler_hdf5.}

\item{score_col}{Column in pixels tibble to plot within the heatmap, defaults to "log10_count".}

\item{file_output}{Output file name. If NULL, no file will be cached.}

\item{fig_width}{Desired width of file. If not provided, will default to 6 (inches implied).}

\item{fig_height}{Desired height of file.}

\item{fig_units}{Unit input for ggsave, defaults to "inches".}

\item{fig_device}{Device input for ggsave, defaults to "png".}

\item{fig_dpi}{DPI input for ggsave, defauts to 300.}

\item{overwrite_cache}{Should the function ignore existing cached images and overwrite them? Defaults to FALSE.}
}
\description{
Executes plot_hic_matrix(), but does so with a file as output. Useful for
caching plot outputs in markdowns, and has the added benefit of configuring
asymmetric matrices them with the correct aspect ratios. To specify
approximate figure size, either height or width can be specified. If both
are provided, these will override the aspect ratio calculation and use these
dimensions directly. Output returns a file name. Note: height/width values
apply to the panel itself, not plot margins or axes. As such, a figure with
<fig_width> of 6 inches will be saved as ~7.5-8 inches wide depending on
labels and legends.
}
