% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_karyotypes.R
\name{plot_karyotypes}
\alias{plot_karyotypes}
\title{Plot karyotypes}
\usage{
plot_karyotypes(
  grange_win = NULL,
  pix_in = NULL,
  pix_axis = "x",
  on_axis = "x",
  break_num = 5,
  axis_position = "top"
)
}
\arguments{
\item{grange_win}{GRanges object to illustrate. Assumes contiguous regions; otherwise use mutliple plot calls.}

\item{pix_in}{Pixel tibble in the format of read_cooler_hdf5() output. If provided, this will be used to determine <grange_win>.}

\item{pix_axis}{Axis of pixel tibble for which to plot karyogram, defaults to "x" (i.e. seqnames1,start1,start2), can also be "y".}

\item{on_axis}{Axis on which to plot karyogram, defaults to "x" and can also be "y".}

\item{break_num}{Number of breaks to split axis into. If multiple chromosomes are included, the start of each chromosome will also be included, so this value is approximate.}

\item{axis_position}{Position of axis; defaults to "top" in which case x-axis is drawn on the top of the panel. If axis is drawn in y dimension, this is changed to "left".}
}
\description{
Plot karyotyping facet for a given genomic range. If no range is provided,
draw for entire genome.
}
