% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombo_matrix_with_kary.R
\name{plotCombo_matrix_with_kary}
\alias{plotCombo_matrix_with_kary}
\title{Arrangement of square matrix panel with karyograms on top and right axes.}
\usage{
plotCombo_matrix_with_kary(p_mtx, p_xkary, p_ykary)
}
\arguments{
\item{p_mtx}{GGplot object for matrix. Axis labels and titles will be extracted from this plot.}

\item{p_xkary}{GGplot object for X axis karyogram.}

\item{p_ykary}{GGplot object for Y axis karyogram.}
}
\description{
Provide a matrix plot and two karyogram plots (one for the X and Y axes),
returns a combined plot with the karyograms inserted.
}
