% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombo_pyramid_tracks.R
\name{plotCombo_pyramid_tracks}
\alias{plotCombo_pyramid_tracks}
\title{Plot HiC matrix (pyramid)}
\usage{
plotCombo_pyramid_tracks(
  mtx_plot,
  kary_plot,
  track_plot = NULL,
  kary_height = 0.05,
  track_height = 0.4,
  show_viewports = TRUE
)
}
\arguments{
\item{gr_input}{GRange region for X-axis. Defaults to entire genome.}

\item{file_cool}{Cool file to plot.}

\item{tb_pixel}{Tibble formatted from call to read_cooler_hdf5(), can include additional columns but must have basics.}

\item{max_pixel_count}{Maximum number of pixels to plot, i.e. max_pixels argument for read_cooler_hdf5.}

\item{silent}{Boolean, defaults to FALSE. Should the function suppress messages about fetching / processing time for read_cooler_hdf5()?}

\item{score_col}{Column in pixels tibble to plot within the heatmap, defaults to "log10_count".}

\item{score_func}{Function to apply to score column, for instance < function(x) log10(x+1) >.}

\item{score_name}{Name to label score with, defaults to the value of <score_col>.}

\item{kary_cm}{Width of karyogram in fraction of plot height/width; defaults to 2cm.}

\item{as_newpage}{Should a new page be started for the plot? Defaults to TRUE.}

\item{sub_tracks}{GGplot with a single panel containing X-axis information.}

\item{sub_track_heigh}{Ratio of plot area that should be dedicated to the X-axis information panel, defaults to 0.3.}

\item{sub_track_x_ax_frac}{Fraction of sub_track plot area to dedicate to that plot's x-axis, defaults to 0.2.}
}
\description{
Basic plot of HiC matrix in a region. Accepts a cooler file name and GRanges
for X and Y axes, but can also accept a tibble (<tb_pixel>) that has been
pre-formatted (generally by a call to read_cooler_hdf5()). In such a case,
GRange and <file_cool> arguments are ignored and their values are inferred
from the input tibble.

NOTE: This function does NOT return standard ggplot objects, it should be
aligned/stored using graphics objects like standard grid and gridExtra
objects.
}
