% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_compartment_file.R
\name{read_compartment_file}
\alias{read_compartment_file}
\title{Read compartment eigenvector file}
\usage{
read_compartment_file(fl_nm)
}
\arguments{
\item{fl_nm}{Filename of eigenvector file to read.}
}
\description{
Read a compartment eigenvector file as output by the hic.cgp
pipeline and return as a granges-format tibble defining merged compartments
throughout the genome. Eigenvectors are reported as averages in each 
compartment along with standard deviations. Note that while compartments are
reduce()'d, empty bins are removed first, so these bins will be missing and
adjacent compartments may be of the same character (i.e., if an A compartment
contains a bin with an NA value for E1, it will be returned as two A
compartments flanking that missing bin).
}
