% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cooler_bins_hdf5.R
\name{read_cooler_bins_hdf5}
\alias{read_cooler_bins_hdf5}
\title{Read cooler genomic bins}
\usage{
read_cooler_bins_hdf5(file_cooler, granges_list = NULL, ignore_strand = FALSE)
}
\arguments{
\item{file_cooler}{Cooler file from which to retrive bin information.}

\item{granges_list}{Genomic ranges to be incorporated. If one or more is provided, bins will be retrieved for each, plus an additional arbitrary bin ID.}

\item{ignore_strand}{Should strand orientation be ignored among input granges?}
}
\description{
Given a cooler file, retrieve genomic bin coordinates and return as a
GRanges object. Does not open file continually, but requires that there not
be an existing connection to another HDF5 based on rhdf5 requirements. If
strand of input GRanges is negative, this is assumed to mean that said
region is in the reverse direction, and so the bin's strand value will be
matched.
}
