% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cooler_hdf5_rot45.R
\name{read_cooler_hdf5_rot45}
\alias{read_cooler_hdf5_rot45}
\title{Read Cooler HDF5, rotate 45 degrees.}
\usage{
read_cooler_hdf5_rot45(
  file_cool,
  gr_range1 = NULL,
  diag_distance = NULL,
  silent = TRUE,
  max_pixels = 6250000,
  cache_tsv = NULL,
  overwrite_cache = FALSE
)
}
\arguments{
\item{file_cool}{Name of .cool file(s) with the appropriate bin size. If multiple files are provided, functions is run recursively and tables are combined (including cache TSV).}

\item{gr_range1}{GRange of contiguous region for X dimension. Defaults to entire genome, and must be contiguous.}

\item{silent}{Should processing time messages be suppressed? Defaults to FALSE.}

\item{max_pixels}{Maximum number of pixels to retrieve without erroring out. Defaults to 6.25 million (2500x2500 grid), can also be set to Inf to disregard.}

\item{overwrite_cache}{Should cache file be overwritten? Defaults to FALSE.}

\item{diag_distace}{Maximum distance from the diagonal (useful for filtering pixels from tables meant for rotated plots). Filter is applied AFTER fetching pixels, so does not generally help performance.}

\item{cache_table}{Filename of TSV to cache pixel results into. If provided and this file exists, this table will be read and returned. If not found, it will be created.}
}
\description{
Read .cool files as HDF5 using the read_cooler_hdf5() function, but perform
required steps to rotate its pixels 45 degrees within a given window. In
short: 1) Read cooler file(s) with read_cooler_hdf5(), 2) rotate pixels with
rotate_pix_45(), 3) trim pixels to the specified X range with
trim_rot45().
}
