% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matrix_file.R
\name{read_matrix_file}
\alias{read_matrix_file}
\title{Read matrix file}
\usage{
read_matrix_file(
  matrix_file_in,
  ranges_x = NULL,
  ranges_y = NULL,
  auto_reflect = TRUE
)
}
\arguments{
\item{matrix_file_in}{Gzipped and tabix-indexed TSV file of matrix values with nine columns.}

\item{ranges_x}{List of GRanges to be included on the X axis.}

\item{ranges_y}{List of GRanges to be included on the Y axis.}

\item{auto_reflect}{Should function be applied across X=Y as well? Defaults to TRUE.}
}
\description{
Function reads tabix-indexed matrix files (TSV format) using
Rsamtools::scanTabix to filter genomic range coordinates. Separate GRanges
can be submitted for asymmetrical matrices (for instance, chr2 on the X axis
vs. chr1 on the Y). Matrix files are tabix-indexed based on their first bin
coordinate ("bin1"), which is arbitrarily treated here as dimension
"X". Files are first filtered using tabix in this dimension, then converted
to GenomicRanges and filtered using subsetByOverlaps(). In effect, all
coordinates that fall within the X dimension are returned by tabix, then
filtered further based on Y coordinates. Theoretically tabix filtering is
faster, but the difference seems negligible in most instances...the chief
advantage with tabix is not reading an entire matrix file into memory.

If both ranges_x and ranges_y are NULL, the entire file is read with
fread(). If only one coordinate is provided, the other is set equal by
default (making a square matrix). For all/one comparisons, submit a GRanges
containing the smaller desired region and another containing all
chromosomes.

To minimize the size of matrix files, if coordinates (i,j) are included then
(j,i) are not as they are functionally equivalent. However, this means that
depending on the sorting algorithms used in the generation of matrix files,
coordinates may be missing if their equivalents are reflected across X=Y (if
you're looking for (i,j) you'll miss it because it's actually (j,i)). To
compensate for this, auto_reflect runs the read_matrix_file() function twice
to compensate, once normally and once with X and Y ranges switched. This can
be disabled, but you should be confident that your selected ranges are all
represented. Auto_reflect is not run if an entire matrix file is read (for
obvious reasons).

Input matrix columns: seqnamesX, startX, endX, seqnamesY, startY, endY, binIDX, binIDY, count.
}
