% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tad_file.R
\name{read_tad_file}
\alias{read_tad_file}
\title{Read TAD file}
\usage{
read_tad_file(
  file_in,
  return_tad_gr = FALSE,
  boundary_column = NULL,
  boundary_calls = 1
)
}
\arguments{
\item{file_in}{Filename to read.}

\item{return_tad_gr}{Should the TAD start/end positions themselves be returned? Defaults to FALSE, in which case each boundary is returned on one row.}

\item{boundary_column}{Should a specific window size be used to determine boundaries (e.g "x5" for the 5x bin size window, etc)? Defaults to NULL.}

\item{boundary_cols}{How many boundary calls should be TRUE for a boundary to be returned? Defaults to 1, in which case at least one of the four window sizes must have been defined as a boundary. Can be set to zero, in which case all pixels are returned.}
}
\description{
Read a TAD boundary file produced by Cooltools via the hic.cgp
pipeline. By default returns a table of boundaries, but can also return a 
GRanges object of tads themselves (i.e. the region between two boundaries).
Boundary calls are reported as TRUE/FALSE in columns that are named in the 
format "x3" to indicate "3 x bin size." For instance, boundary calls measured
with in a 100kb resolution matrix with a window size of 300k will be reported
in the column "x3," as will a 10kb matrix with a 30kb window size.
}
