% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect_pixel_matrix.R
\name{reflect_pixel_matrix}
\alias{reflect_pixel_matrix}
\title{Reflect pixel matrix.}
\usage{
reflect_pixel_matrix(tb_pix_in, drop_diagonal = FALSE)
}
\arguments{
\item{tb_pix_in}{Tibble of pixel data (as output by read_cooler_hdf5(), typically).}

\item{drop_diagonal}{Boolean, should diagonal points (bin1==bin2) be omitted? Defaults to FALSE.}
}
\description{
Reflect a (non-rotated) pixel matrix across the diagonal (i.e. swap columns
with start/end/seqnames values). Note that the only columns that are swapped
are the seqnames, start, end, start_adj, end_adj, and bin1/bin2 columns (If
present in the data, bin_alt1/2 columns will be swapped as well). Others will
need to be swapped manually, for instance via:

swap_columns(tb_pix,"column1","column2")

To use rotate_pix_45(), perform the reflection first, i.e.:

read_cooler_hdf5(<file_in>) %>%
 reflect_pixel_matrix %>%
 rotate_pix_45
}
