% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_pix_45.R
\name{rotate_pix_45}
\alias{rotate_pix_45}
\title{Rotate pixels 45 degrees.}
\usage{
rotate_pix_45(df_in)
}
\arguments{
\item{df_in}{Pixel tibble, generally the output format of read_cooler_hdf5().}
}
\description{
Given a table of pixel data, append 'x_coords' and 'y_coords'
columns for X and Y values rotated 45 degrees about their center. Useful
for plotting HiC data along a horizontal axis without manually rotating
plots. Output tables are in long format with duplicated values (each pixel
needs four rows, one for each pair of (x,y) coordinates). Adds a "pix_id"
column, as well, so ggplot aesthetics should look similar to:

ggplot(tb_pix,aes(x=x_coords,y=y_coords,group=pix_id,fill=log10_counts))

NOTE: If color/fill values appear to be broken, make sure "group=pix_id" has
been specified.

While the resulting X-axis should be considerably longer than that of the
input plot (2 x resolution^2 = c^2, so pixel width should be
sqrt(2 x res^2)), coordinates are divided by sqrt(2) such that rotated pixel
diagonals are equal to original pixel sides. This occurs for both X and Y
dimensions, and should be consistent for any condition in which matrix is
centered on the diagonal.
}
