% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_rot45.R
\name{trim_rot45}
\alias{trim_rot45}
\title{Trim rotated pixels}
\usage{
trim_rot45(tb_in, gr_x = NULL, x_range = NULL, y_range = NULL)
}
\arguments{
\item{tb_in}{Tibble input, generally output of rotate_pix_45().}

\item{gr_x}{If provided, all pixels will be filtered to keep those which fall inside this genomic region.}

\item{x_range}{Vector of upper and lower bounds to use manually on x axis, i.e. the "start_adj1/end_adj1" boundaries to filter by. Ignored if gr_x is provided.}

\item{y_range}{Vector of upper and lower bounds to use manually on y axis, i.e. the "start_adj2/end_adj2" boundaries to filter by.}
}
\description{
Given a tibble with rotated pixels (output of rotate_pix_45()),
perform two functions. First, if a grange object has been provided as
<gr_x>, remove all rows where the x-axis values are outside of this range.
Second, remove all rows which would fall outside of the x and y axes in
in general. This has the effect of "shaving" top,bottom,and side "corners"
from each pixel such that the top/bottom/side boundaries are flat, i.e. the
pixels become triangles. Looks cleaner in plots, plus GGplot's clipping
system doesn't like to draw partial geoms. If no arguments are provided
besides the input tibble, edge pixels are determined and trimmed
appropriately.
}
