% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_score_column.R
\name{normalize_score_column}
\alias{normalize_score_column}
\title{Normalize HiC score column}
\usage{
normalize_score_column(
  tb_pix,
  score_column = "log10_count_wgs_nrm",
  new_col_name = NULL
)
}
\arguments{
\item{tb_pix}{Pixel tibble to normalize.}

\item{score_column}{Score column to normalize, defaults to "log10_count_wgs_nrm".}

\item{new_col_name}{Name for normalized column. If NULL, original column will be replaced with normalized values.}
}
\description{
When plotting multiple HiC interaction matrices together, score
can (should?) be normalized based on reads in each matrix. This is best
accomplished by taking the sum of score in both samples and dividing each by
the lowest sum. Score columns are then divided by their respective ratio.
For instance, samples with 1000 and 2000 reads each will have ratios of 1
and 2, respectively; score columns are then divided by 1 and 2, ultimately
reducing the scores in sample 2 by half.
}
