% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_cna.R
\name{patchwork_cna}
\alias{patchwork_cna}
\title{Patchwork CNA}
\usage{
patchwork_cna(regions_tb, bdg_file, score_fill, num_y_brks = 3)
}
\arguments{
\item{regions_tb}{Region table, generally the output of patchwork_bins(boundaries_only=TRUE).}

\item{bdg_file}{Filename of Tabix-indexed file. Assumes index file is in place.}

\item{score_fill}{Fill color of rectangles to draw. If not provided, a fill based on log2(CNA raw) is applied.}

\item{num_y_brks}{Number of Y-axis breaks to be drawn. Defaults to 3.}
}
\description{
Plot raw CNA values. Returns a GGplot object.
}
